requirements.txt
Flask==3.0.3
openai==1.35.13
requests==2.32.3
beautifulsoup4==4.12.3
psycopg2-binary==2.9.9
python-dotenv==1.0.1
google-generativeai==0.7.0 